misc debugs and add conus reservoir da test

+++ test/nhf/conus/test_conus.py
@@ -13,15 +13,12 @@ from ..utils.make_configs import Config
 from ..utils.make_forcing import build_forcing_dataset
 
 START_TIME = "2000-01-01 00:00"
-END_TIME = "2000-01-01 12:00"
+END_TIME = "2000-01-01 04:00"
 FORCING_MODE = "constant"
 CONSTANT_QLAT = 10
 
 DATA_DIR = Path(__file__).parent / "data" / "base"
 CFG = Config(DATA_DIR, START_TIME, END_TIME, max_loop_size=1)
-
-CFGS = []
-
 
 def setup(source_gpkg: str | Path, refresh: bool = True):
     """Subset the NHF domain and generate forcing for a standard test case."""
@@ -29,7 +26,8 @@ def setup(source_gpkg: str | Path, refresh: bool = True):
         CFG.write_yaml()
 
     if refresh or not CFG.domain_path.exists():
-        CFG.domain_path.unlink()
+        if CFG.domain_path.exists():
+            CFG.domain_path.unlink()
         CFG.domain_path.symlink_to(source_gpkg)
 
     if refresh or not has_files(CFG.channel_forcing_dir, CFG.qlat_file_pattern):
@@ -51,13 +49,3 @@ def test_conus():
     assert_output_dimensions_and_validity(
         CFG.output_dir, CFG.domain_path, CFG.config_path
     )
-
-
-@pytest.mark.integration
-def test_conus_reservoir_da():
-    pass
-
-
-@pytest.mark.integration
-def test_conus_reservoir_coarse():
-    pass
